\name{getCounter}
\alias{getCounter}
\title{Returns the counter associated with a table, figure, or equation}
\usage{
getCounter(
    type = c("table", "figure", "equation")
  )
}
\arguments{
  \item{type}{Character string indicating either a \sQuote{table}, \sQuote{figure}, or \sQuote{equation}.}
}
\description{
Returns the counter associated with a table, figure, or equation
}
