\name{Literasee-package}
\alias{Literasee-package}
\alias{Literasee}
\docType{package}
\title{Literasee: Assessment Literacy Content Development via R}
\description{
Assessment Literacy Content Development via R.
}

\details{
\tabular{ll}{
Package: \tab Literasee\cr
Type: \tab Package\cr
Version: \tab 0.0-1.0\cr
Date: \tab 2017-5-26\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
Literasee is an experimental package for the R software environment designed to ease in the creation of assessment literacy related content.
}
\author{
Damian W. Betebenner \email{dbetebenner@nciea.org} and Adam Van Iwaarden \email{avaniwaarden@gmail.com}
}

\references{
Betebenner, D. W. (2009). Norm- and criterion-referenced student growth.
Educational Measurement: Issues and Practice, 28(4):42-51.
}
