\name{numbers2words}
\alias{numbers2words}
\title{Returns the counter associated with a table, figure, or equation}
\usage{
numbers2words(
    x=0L
  )
}
\arguments{
  \item{x}{Integer to be converted to (English) word(s). Default is 0 (zero)}
}
\description{
Taken from the internet. See \href{https://stackoverflow.com/a/46652141}{this Stack Overflow conversation} and \href{https://gist.githubusercontent.com/hack-r/22104543e2151519c41a8f0ce042b31c/raw/01731f3176b4ee471785639533d38eda4790ab77/numbers2words.r}{Gist solution}.
}
