\name{reportSGP}
\alias{reportSGP}
\title{Function for automatic technical report generation from SGP analyses.}
\description{
The function reportSGP is used to create automatic or customized technical reports associated with SGP analyses using the SGP object produced as part of the SGP analysis.
}

\usage{
reportSGP(sgp_object=NULL)
}

\arguments{
  \item{sgp_object}{An SGP object resulting from SGP analyses.
  }
}

\value{Returns a technical in multiple output formats including HTML and PDF.
}

\author{Damian W. Betebenner \email{dbetebenner@nciea.org} and Adam Van Iwaarden \email{avaniwaarden@gmail.com}}

\examples{
    reportSGP()
}

\keyword{documentation}
