\name{Literasee-package}
\alias{Literasee-package}
\alias{Literasee}
\docType{package}
\title{Literasee: Assessment Literacy Content Development via R}
\description{
Assessment Literacy Content Development via R.
}

\details{
\tabular{ll}{
Package: \tab Literasee\cr
Type: \tab Package\cr
Version: \tab 0.0-5.2\cr
Date: \tab 2020-3-29\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
Literasee is an experimental package for the R software environment designed to help create assessment literacy related content.
}
\author{
Adam R. Van Iwaarden \email{avaniwaarden@nciea.org} and Damian W. Betebenner \email{dbetebenner@nciea.org}
}

\references{
Betebenner, D. W. (2009). Norm- and criterion-referenced student growth.
Educational Measurement: Issues and Practice, 28(4):42-51.
}
